[multi-lora] operations: fold the round executor into operation_backend
The driver-side helpers cohabit with the actor-side class the same way
controller.py keeps create_multilora_controller next to the actor; one
mechanism, one file, and the module stays importable without megatron so
round-logic tests keep running on CPU.
