+++ train_multi_lora_async.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import ray
 
 from miles.ray.multi_lora.controller import create_multilora_controller, get_multi_lora_controller
-from miles.ray.multi_lora.operation_round import apply_tinker_defaults, run_operation_round
+from miles.ray.multi_lora.operation_backend import apply_tinker_defaults, run_operation_round
 from miles.ray.placement_group import create_placement_groups, create_rollout_manager, create_training_models
 from miles.utils import object_store
 from miles.utils.adapter_config import parse_adapter_run_yaml
